// src/visitas/enums/visita.enums.ts
export enum VisitaEstado {
  PENDIENTE = 'PENDIENTE',
  EN_CURSO = 'EN_CURSO',
  COMPLETADA = 'COMPLETADA',
  CANCELADA = 'CANCELADA',
}

export enum EventoVisitaTipo {
  CREADA = 'CREADA',
  REPROGRAMADA = 'REPROGRAMADA',
  ASIGNADA_TECNICO = 'ASIGNADA_TECNICO',
  ASIGNADA_SUPERVISOR = 'ASIGNADA_SUPERVISOR',
  CHECK_IN = 'CHECK_IN',
  CHECK_OUT = 'CHECK_OUT',
  CANCELADA = 'CANCELADA',
  NOTA_SUP = 'NOTA_SUP',
  NOTA_TEC = 'NOTA_TEC',
  COMPLETADA = 'COMPLETADA',
}

export enum EvidenciaTipo {
  FOTO = 'FOTO',
  VIDEO = 'VIDEO',
  AUDIO = 'AUDIO',
  DOCUMENTO = 'DOCUMENTO',
  FIRMA = 'FIRMA',
}